# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeepRef API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ingestions": {
      "get": {
        "tags": [
          "ingestions"
        ],
        "operationId": "listIngestions",
        "responses": {
          "200": {
            "description": "Ingestions ordered newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestionDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingestions"
        ],
        "operationId": "createIngestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Ingestion created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ingestion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingestions/{ingestion_id}": {
      "get": {
        "tags": [
          "ingestions"
        ],
        "operationId": "getIngestion",
        "parameters": [
          {
            "name": "ingestion_id",
            "in": "path",
            "description": "Ingestion identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionDto"
                }
              }
            }
          },
          "404": {
            "description": "Ingestion not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingestions/{ingestion_id}/cancel": {
      "post": {
        "tags": [
          "ingestions"
        ],
        "operationId": "cancelIngestion",
        "parameters": [
          {
            "name": "ingestion_id",
            "in": "path",
            "description": "Ingestion identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Cancellation accepted"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingestions/{ingestion_id}/items": {
      "get": {
        "tags": [
          "ingestions"
        ],
        "operationId": "listIngestionItems",
        "parameters": [
          {
            "name": "ingestion_id",
            "in": "path",
            "description": "Ingestion identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestionItemDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIngestion": {
        "type": "object",
        "required": [
          "project_id",
          "seed_dois"
        ],
        "properties": {
          "citation_provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_depth": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "metadata_provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "seed_dois": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IngestionDto": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "max_depth",
          "seed_count",
          "queued_count",
          "fetched_count",
          "failed_count",
          "created_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "failed_count": {
            "type": "integer",
            "format": "int32"
          },
          "fetched_count": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "max_depth": {
            "type": "integer",
            "format": "int32"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "queued_count": {
            "type": "integer",
            "format": "int32"
          },
          "seed_count": {
            "type": "integer",
            "format": "int32"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "IngestionItemDto": {
        "type": "object",
        "required": [
          "doi",
          "depth",
          "status",
          "attempts",
          "queued_at"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "doi": {
            "type": "string"
          },
          "fetched_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_doi": {
            "type": [
              "string",
              "null"
            ]
          },
          "queued_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}